+++ src/main/kotlin/io/unthrottled/doki/icon/DokiIcons.kt
@@ -14,6 +14,10 @@ object DokiIcons {
       val TOOL_WINDOW = load("icons/plugins/motivator/motivator_toolwindow.svg")
     }
 
+    object Icons {
+      val TOOL_WINDOW = load("icons/doki/neko_glyph.svg")
+    }
+
     object AMII {
       val TOOL_WINDOW = load("icons/plugins/amii/plugin-tool-window.svg")
     }