+++ changelog/CHANGELOG.md
@@ -1,6 +1,18 @@
 Changelog
 ---
 
+# 88.2-1.6.0 [Doki Theme Icons]
+
+<div align="center">
+<img src="https://doki.assets.unthrottled.io/misc/doki_icons_logo.png" width="256" height="256" alt="Doki Theme Icons"></img>
+</div>
+
+- Removed the legacy Material Icons. Please install the [Doki Theme Icons](https://github.com/doki-theme/doki-theme-icons-jetbrains) in its place. 
+- Renamed plugin to `Doki Theme` & updated plugin logo.
+- Zero Two's Obsidian theme is now the default for new users.
+- Updated Chocola's Xmas theme.
+- Made Natsuki's & Tohru's ANSI green color more usable.
+
 # 88.1-1.5.4 [Better Icon Loading]
 
 - Potentially fixed icon loading issue at start up.